# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Resource OpenApi",
        "contact": {
            "name": "学院君",
            "url": "http://xueyuanjun.com",
            "email": "support@todo.test"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://todo.test/dingoapi/tasks"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks",
                "description": "Returns list of tasks",
                "operationId": "getTaskList",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept header to specify api version",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page num of the list",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The item num per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The result of tasks"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your email / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {}
                    }
                },
                "scheme": "http"
            }
        }
    },
    "security": [
        []
    ]
}